@@ -1,13 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import React from "react"
+import ReactDOM from "react-dom/client"
+import "./index.css"
+import App from "./App"
+import MenuContextProvider from "./context/MenuContextProvider"
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
-
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(
+	<React.StrictMode>
+		<MenuContextProvider>
+			<App />
+		</MenuContextProvider>
+	</React.StrictMode>
+)
