@@ -1,24 +1,16 @@
-import Navbar from "./components/Navbar"
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
-import Home from "./pages/Home"
-import Projects from "./pages/Projects"
-import Story from "./pages/Story"
-import Error404 from "./pages/Error404"
-import Footer from "./components/Footer"
+import { Error404Page, HomePage, ProjectsPage} from "./pages"
 
 function App() {
 	return (
 		<>
 			<Router>
-				<Navbar />
 				<Routes>
-					<Route path='/'  element={<Home/>}/>
-					<Route path='/Projects' element={<Projects/>}/>
-					<Route path='/Blog'>Blog</Route>
-					<Route path='/story' element={<Story/>}/>	
-					<Route path='/*' element={<Error404/>}/>	
+					<Route path='/' element={<HomePage />} />
+					<Route path='/projects' element={<ProjectsPage />} />
+					{/* <Route path='/story' element={<StoryPage />} /> */}
+					<Route path='*' element={<Error404Page />} />
 				</Routes>
-				<Footer/>
 			</Router>
 		</>
 	)
